backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
python-dotenv
pydantic
